+++ src/error.rs
@@ -4,12 +4,6 @@ use thin_vec::{thin_vec, ThinVec};
 
 use crate::ErrorKind;
 
-/// This is boxed inside an `Error` to make sure that all function signatures
-/// involving it aren't inflated. See `Error` for helper functions.
-pub struct ErrorInner {
-    pub stack: ThinVec<(ErrorKind, Option<&'static Location<'static>>)>,
-}
-
 /// An experimental error struct that has an internal stack for different kinds
 /// of errors and a stack for locations. This is a replacement for the bad
 /// information you get from backtraces within `async` tasks.
@@ -21,14 +15,19 @@ pub struct ErrorInner {
 ///
 /// Use at least `.map_add_err(|| ())` before every time an error is propogated
 /// up the stack to make sure the location stack is filled.
-pub struct Error(pub ErrorInner);
+pub struct Error {
+    // using a ThinVec has many advantages from taking as little space as possible, having single
+    // indirection vs. other methods, and having the niche optimizations applied to `Result<(),
+    // Error>` and others.
+    pub stack: ThinVec<(ErrorKind, Option<&'static Location<'static>>)>,
+}
 
 impl Debug for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         // in reverse order of a typical stack, I don't want to have to scroll up to see
         // the more specific errors
         f.write_fmt(format_args!("Error {{ stack: [\n"))?;
-        for (i, (error, location)) in self.0.stack.iter().enumerate().rev() {
+        for (i, (error, location)) in self.stack.iter().enumerate().rev() {
             match error {
                 ErrorKind::UnitError => (),
                 ErrorKind::StrError(s) => {
@@ -58,13 +57,19 @@ impl Debug for Error {
 }
 
 impl Error {
+    pub fn empty() -> Self {
+        Self {
+            stack: ThinVec::new(),
+        }
+    }
+
     /// Use `MapAddErr` instead of this
     #[track_caller]
     pub fn from_kind<K: Into<ErrorKind>>(kind: K) -> Self {
         let l = Location::caller();
-        Self(ErrorInner {
+        Self {
             stack: thin_vec![(kind.into(), Some(l))],
-        })
+        }
     }
 
     /// Returns a base timeout error
@@ -84,7 +89,7 @@ impl Error {
 
     /// The same as [Error::add_err] but without pushing location to stack
     pub fn add_err_no_location<K: Into<ErrorKind>>(mut self, kind: K) -> Self {
-        self.0.stack.push((kind.into(), None));
+        self.stack.push((kind.into(), None));
         self
     }
 
@@ -93,22 +98,21 @@ impl Error {
     /// `ok_or_else`.
     #[track_caller]
     pub fn add_err<K: Into<ErrorKind>>(mut self, kind: K) -> Self {
-        self.0.stack.push((kind.into(), Some(Location::caller())));
+        self.stack.push((kind.into(), Some(Location::caller())));
         self
     }
 
     /// Only adds `track_caller` location to the stack
     #[track_caller]
     pub fn add_location(mut self) -> Self {
-        self.0
-            .stack
+        self.stack
             .push((ErrorKind::UnitError, Some(Location::caller())));
         self
     }
 
     /// Returns if a `TimeoutError` is in the error stack
     pub fn is_timeout(&self) -> bool {
-        for (error, _) in &self.0.stack {
+        for (error, _) in &self.stack {
             if matches!(error, ErrorKind::TimeoutError) {
                 return true
             }
@@ -118,7 +122,7 @@ impl Error {
 
     /// Chains the stacks of `other` onto `self`
     pub fn chain_errors(mut self, mut other: Self) -> Self {
-        self.0.stack.append(&mut other.0.stack);
+        self.stack.append(&mut other.stack);
         self
     }
 }