+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changes
  - Added  `Send + Sync` bounds to the `BoxedError` variant so that the whole `Error` struct is now `Send + Sync`
  - Used `ThinVec` to reduce the size of `Result<(), Error>` to be the same as `usize`
+ - Refactored the public internals of `Error`
 
 ## [0.2.0] - 06-06-2023
 ### Additions